+++ packages/eth-rpc-client/.npmignore
