+++ packages/eth-rpc-client/src/globals.d.ts
